requirements.txt
Flask==1.1.4
PyMySQL==0.10.1
cryptography==3.4.8
python-dotenv==0.19.2

